+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project are documented in this file.
 
 
 ## 2025-10-25
+### Changes
+- Login now directs to first Workspace 
 ### Removed
 - Ability to navigate backwards from TodoLists page